GeneSearch/requirements.txt
Package               Version
--------------------- -------
asgiref               3.5.2
Django                4.1
django-rest-framework 0.1.0
djangorestframework   3.13.1
pip                   22.2.2
pytz                  2022.1
setuptools            58.1.0
sqlparse              0.4.2
typing_extensions     4.3.0
